pyproject.toml
[project]
name = "apikey-config"
version = "0.1.0"
description = "Secure API key manager with public key verification"
authors = [{name = "Cheng Yanru", email = "yanru@cyanru.com"}]
requires-python = ">=3.8"

dependencies = [
    "argon2-cffi>=23.1.0",
]

[project.scripts]
apikey = "apikey:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
